Less warnings

+++ src/common/widget/column.hpp
@@ -1,13 +1,13 @@
 #ifndef _TMS_COMMON_WIDGET__COLUMN_HPP_
 #define _TMS_COMMON_WIDGET__COLUMN_HPP_
 
-#include <cstdio>
+#include <string>
 
 struct Column {
-  size_t field_id;
+  int field_id;
   std::string name;
-  size_t width;
-  Column(size_t field_id, std::string name, size_t width) :
+  int width;
+  Column(int field_id, std::string name, int width) :
     field_id(field_id), name(name), width(width) {}
 };
